+++ src/environments/globals.dev.ts
@@ -9,8 +9,6 @@ export const debug = false;
 
 // export const base_path = "http://nimble-staging.salzburgresearch.at";
 // export const ub_base = "http://nimble-staging.salzburgresearch.at/search";
-// export const simple_search_endpoint="http://nimble-staging.salzburgresearch.at/marmotta/solr/catalogue_multilang";
-// export const simple_search_properties_endpoint="http://nimble-staging.salzburgresearch.at/marmotta/solr/prop_trial";
 // export const pw_reset_link = "http://nimble-staging.salzburgresearch.at:8080/auth/realms/master/login-actions/reset-credentials?client_id=nimble_client";
 
 
@@ -18,8 +16,6 @@ export const debug = false;
 
 // export const base_path = "https://nimble-platform.salzburgresearch.at/nimble";
 // export const ub_base = "https://nimble-platform.salzburgresearch.at/ub-search";
-// export const simple_search_endpoint="https://nimble-platform.salzburgresearch.at/marmotta/solr/catalogue_multilang";
-// export const simple_search_properties_endpoint="https://nimble-platform.salzburgresearch.at/marmotta/solr/prop_trial";
 // export const pw_reset_link = "https://nimble-platform.salzburgresearch.at:8080/auth/realms/master/login-actions/reset-credentials?client_id=nimble_client";
 
 
@@ -27,8 +23,6 @@ export const debug = false;
 
 // export const base_path = "https://nimble.salzburgresearch.at";
 // export const ub_base = "https://hydra2.ikap.biba.uni-bremen.de:8443";
-// export const simple_search_endpoint="https://nimble-platform.salzburgresearch.at/marmotta/solr/catalogue_multilang";
-// export const simple_search_properties_endpoint="https://nimble-platform.salzburgresearch.at/marmotta/solr/prop_trial";
 // export const pw_reset_link = "https://nimble-platform.salzburgresearch.at:8080/auth/realms/master/login-actions/reset-credentials?client_id=nimble_client";
 
 
@@ -36,8 +30,6 @@ export const debug = false;
 
 // export const base_path = "http://nimble.uk-south.containers.mybluemix.net";
 // export const ub_base = "https://hydra2.ikap.biba.uni-bremen.de:8443";
-// export const simple_search_endpoint="https://nimble-platform.salzburgresearch.at/marmotta/solr/catalogue_multilang";
-// export const simple_search_properties_endpoint="https://nimble-platform.salzburgresearch.at/marmotta/solr/prop_trial";
 // export const pw_reset_link = "https://nimble-platform.salzburgresearch.at:8080/auth/realms/master/login-actions/reset-credentials?client_id=nimble_client";
 
 
@@ -45,8 +37,6 @@ export const debug = false;
 
 export const base_path = "http://localhost";
 export const ub_base = "http://localhost:8090";
-export const simple_search_endpoint="http://localhost:8082/solr/catalogue_multilang";
-export const simple_search_properties_endpoint="http://localhost:8082/solr/prop_trial";
 export const pw_reset_link = "http://localhost:8080/auth/realms/master/login-actions/reset-credentials?client_id=nimble_client";
 
 
@@ -58,6 +48,7 @@ export const bpe_endpoint=`${base_path}/business-process`;
 export const data_channel_endpoint=`${base_path}/data-channel`;
 export const data_aggregation_endpoint=`${base_path}/data-aggregation`;
 export const trust_service_endpoint=`${base_path}/trust`;
+export const indexing_service_endpoint=`${base_path}/index`;
 
 
 // BIBA endpoints
@@ -89,7 +80,7 @@ export const config = {
     "eClass": {
       "hiddenCategories": [],
       "logisticsCategory": "14000000",
-      "ontologyPrefix": "http://www.nimble-project.org/resource/eclass/"
+      "ontologyPrefix": "http://www.nimble-project.org/resource/eclass#"
     },
     "FurnitureOntology": {
       "hiddenCategories": ["Catalogue","Company","ContactPerson","Guarantee","Price","Process","Standard","Style","Technique"],
@@ -163,42 +154,30 @@ export const config = {
 
 // Catalogue format variables
 
-export const product_name = "item_name";
-export const product_vendor_id = "item_manufacturer_id";
-export const product_vendor_name = "item_manufacturer_name";
-export const product_vendor_rating = "trust_company_rating";
-export const product_vendor_rating_seller = "trust_company_seller_communication";
-export const product_vendor_rating_fulfillment = "trust_company_fulfillment_of_terms";
-export const product_vendor_rating_delivery = "trust_company_delivery_packaging";
-export const product_vendor_trust = "trust_company_trust_score";
-export const product_img = "item_image";
-export const product_price = "item_price";
-export const product_currency = "item_price_currency";
-export const product_cat = "item_commodity_classification";
-export const product_cat_mix = "item_commodity_classification_mix";
-export const product_filter_prod = ["item_free_of_charge_indicator","item_certificate","item_applicable_countries"];
-export const product_filter_comp = ["item_manufacturer_name","item_manufacturer_origin","manufacturer_certificate","manufacturer_ppap_compliance_level","manufacturer_ppap_document"];
-export const product_filter_trust = ["trust_company_trust_score","trust_company_rating","trust_company_seller_communication","trust_company_fulfillment_of_terms","trust_company_delivery_packaging"];
+export const product_vendor = "manufacturer";
+export const product_vendor_id = "id";
+export const product_vendor_name = "legalName";
+export const product_vendor_rating = "trustRating";
+export const product_vendor_rating_seller = "trustSellerCommunication";
+export const product_vendor_rating_fulfillment = "trustFullfillmentOfTerms";
+export const product_vendor_rating_delivery = "trustDeliveryPackaging";
+export const product_vendor_trust = "trustScore";
+export const product_name = "label";
+export const product_description = "description";
+export const product_img = "imgageUri";
+export const product_price = "price";
+export const product_currency = "currency";
+export const product_cat = "classificationUri";
+export const product_cat_mix = "commodityClassficationUri";
+export const product_filter_prod = ["freeOfCharge","certificateType","applicableCountries"];
+export const product_filter_comp = ["manufacturer.legalName","manufacturer.origin","manufacturer.certificateType","manufacturer.ppapComplianceLevel","manufacturer.ppapDocumentType"];
+export const product_filter_trust = ["manufacturer.trustScore","manufacturer.trustRating","manufacturer.trustSellerCommunication","manufacturer.trustFullfillmentOfTerms","manufacturer.trustDeliveryPackaging","manufacturer.trustNumberOfTransactions"];
 export const product_filter_mappings = {
-  "item_price": "Price",
-  "item_price_currency": "Currency",
-  "item_free_of_charge_indicator": "Free Sample",
-  "item_certificate": "Product / Service Certificate",
-  "item_applicable_countries": "Destination Country",
-  "item_manufacturer_name": "Vendor Name",
-  "item_manufacturer_origin": "Vendor Origin",
-  "manufacturer_certificate": "Vendor Certificate",
-  "manufacturer_ppap_compliance_level": "PPAP Compliance Level",
-  "manufacturer_ppap_document": "PPAP Document",
-  "trust_company_trust_score": "Trust Score",
-  "trust_company_rating": "Overall Rating",
-  "trust_company_seller_communication": "Seller Communication",
-  "trust_company_fulfillment_of_terms": "Fulfillment of Terms",
-  "trust_company_delivery_packaging": "Delivery & Packaging"
+  "price": "Price",
+  "currency": "Currency"
 };
-export const product_nonfilter_full = ["id","_version_","item_description","item_price","item_price_currency","item_commodity_classification_code","item_commodity_classification_uri","item_properties"];
-export const product_nonfilter_regex = ["lmf.","_id"];
+export const product_nonfilter_full = ["_text_","_version_","id","image","localName","languages","catalogueId","doctype","manufacturerId","manufacturerItemId"];
+export const product_nonfilter_regex = ["lmf.","_id", "_txt", "_desc", "_label", "_key", "_price", "_currency", "httpwwwnimbleprojectorgresourceeclasshttpwwwnimbleprojectorgresourceeclasshttpwwwnimbleprojectorgresourceeclasshttpwwwnimbleprojectorgresourceeclass"];
 export const product_configurable = [];
 export const product_default = {};
 export const facet_min = 1;
-export const negotiatables = ["size", "duration"];